refactor(tangle-dapp): Move liquid staking types to shared ui lib
