+++ apps/tangle-dapp/data/liquidStaking/useLsValidatorSelectionTableColumns.tsx
@@ -6,6 +6,12 @@ import {
 } from '@tanstack/react-table';
 import { LiquidStakingItem } from '@webb-tools/tangle-shared-ui/types/liquidStaking';
 import {
+  Collator,
+  Dapp,
+  PhalaVaultOrStakePool,
+  Validator,
+} from '@webb-tools/tangle-shared-ui/types/liquidStaking';
+import {
   Avatar,
   CheckBox,
   Chip,
@@ -19,12 +25,6 @@ import { twMerge } from 'tailwind-merge';
 
 import { StakingItemExternalLinkButton } from '../../components/LiquidStaking/StakingItemExternalLinkButton';
 import TokenAmountCell from '../../components/tableCells/TokenAmountCell';
-import {
-  Collator,
-  Dapp,
-  PhalaVaultOrStakePool,
-  Validator,
-} from '../../types/liquidStaking';
 import calculateCommission from '../../utils/calculateCommission';
 import formatBn from '../../utils/formatBn';
 import formatPercentage from '../../utils/formatPercentage';
