+++ apps/tangle-dapp/data/liquidStaking/useLsValidators.ts
@@ -1,16 +1,16 @@
 import { BN_ZERO } from '@polkadot/util';
 import { LsProtocolId } from '@webb-tools/tangle-shared-ui/types/liquidStaking';
 import { LiquidStakingItem } from '@webb-tools/tangle-shared-ui/types/liquidStaking';
-import { useCallback, useEffect, useMemo, useState } from 'react';
-
-import { LsNetworkId } from '../../constants/liquidStaking/types';
-import useLocalStorage, { LocalStorageKey } from '../../hooks/useLocalStorage';
 import {
   Collator,
   Dapp,
   PhalaVaultOrStakePool,
   Validator,
-} from '../../types/liquidStaking';
+} from '@webb-tools/tangle-shared-ui/types/liquidStaking';
+import { useCallback, useEffect, useMemo, useState } from 'react';
+
+import { LsNetworkId } from '../../constants/liquidStaking/types';
+import useLocalStorage, { LocalStorageKey } from '../../hooks/useLocalStorage';
 import getLsProtocolDef from '../../utils/liquidStaking/getLsProtocolDef';
 import {
   fetchChainDecimals,
