+++ apps/tangle-dapp/types/liquidStaking.ts
@@ -1,52 +1,9 @@
-import { BN } from '@polkadot/util';
 import {
-  LiquidStakingItem,
-  LsProtocolId,
+  Collator,
+  Dapp,
+  PhalaVaultOrStakePool,
+  Validator,
 } from '@webb-tools/tangle-shared-ui/types/liquidStaking';
-
-// All chains
-export type StakingItem = {
-  id: string; // address - Validator, contract address - DAPP, pool/vault ID - VaultOrStakePool
-  totalValueStaked: BN;
-  minimumStake?: BN;
-  chainId: LsProtocolId;
-  chainDecimals: number;
-  chainTokenSymbol: string;
-  itemType: LiquidStakingItem;
-  href: string;
-};
-
-// Chains - Polkadot, Kusama, Tangle etc.
-export type Validator = {
-  validatorAddress: string;
-  validatorIdentity: string;
-  validatorCommission: BN;
-  validatorAPY?: number;
-} & StakingItem;
-
-// Chain - Phala Network (Stake on Vaults or Stake Pools)
-export type PhalaVaultOrStakePool = {
-  vaultOrStakePoolID: string;
-  vaultOrStakePoolName: string;
-  vaultOrStakePoolAccountID: string;
-  commission: BN;
-  type: string;
-} & StakingItem;
-
-// Chain - Astar (dApp Staking)
-export type Dapp = {
-  dappContractAddress: string;
-  dappName: string;
-  dappContractType: string;
-  commission: BN;
-} & StakingItem;
-
-// Chains - Moonbeam, Manta
-export type Collator = {
-  collatorAddress: string;
-  collatorIdentity: string;
-  collatorDelegationCount: number;
-} & StakingItem;
 
 export type LiquidStakingItemType =
   | Validator
